+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## <next>
 * New `info` cell message type and action for setting multiple messages.
+* Update column settings component to utilize `react-select-order-list`
 
 ## 3.2.0
 * Add checkbox to select all columns in column settings